+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 * `insar_tops_burst` workflow now takes the burst product names rather than the SLC product names.
 * Replace `--range-looks` and `--azimuth-looks` options with a single `--looks` option for `insar_tops` and `insar_tops_burst` workflows
+* `insar_tops_burst` workflow now always uses the oldest granule as the reference.
 
 ## [0.3.0]
 ### Added